# Backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BeetleX API",
    "description": "API documentation for the BeetleX Hackathon Platform Backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/events/{id}/registrations": {
      "get": {
        "summary": "List registrations for an event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string" }, "description": "Set to 'csv' to export" }
        ],
        "responses": {
          "200": { "description": "Registrations list retrieved" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}